Add support for indexer hall effect sensors
This commit adds getters to the Shooter subsytem which get the state
of the hall effect sensors. These need to be tested and the logic may need
to be reversed (ie. A non tripped hall effect sensor may return "true")
These getters are called isIndexerSensed(Up/Down)().

This commit also changes the shooting sequence to use these sensors, as well
as adds a much larger timeout to indexing actions so it will be fairly obvious when
things are breaking.